s/requirements.txt
pydantic>=2.0.0
pydantic-ai>=0.7.0
python-dotenv>=1.0.0
colorlog>=6.0.0
typer>=0.9.0
rich>=13.0.0
httpx>=0.25.0
